add test for components
