+++ package.json
@@ -42,6 +42,7 @@
     "recompose": "0.30.0",
     "redux": "4.0.4",
     "redux-devtools-extension": "2.13.8",
-    "redux-thunk": "2.3.0"
+    "redux-thunk": "2.3.0",
+    "reselect": "4.0.0"
   }
 }
