+++ src/components/city-list/city-list.jsx
@@ -2,16 +2,16 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
 
-import {ActionCreator} from '../../reducer/reducer';
+import ActionCreator from '../../action-creator/action-creator';
 
 const CityList = (props) => {
-  const {city, allOffers, cities, changeCityClickHandler} = props;
+  const {city, offers, cities, changeCityClickHandler} = props;
   return <ul className="locations__list tabs__list">
     {[...new Set(cities)].map((offerCity) => <li className="locations__item" key={`location-${offerCity}`}>
       <a
         className={`locations__item-link tabs__item ${city === offerCity && `tabs__item--active`}`}
         href="#"
-        onClick={() => changeCityClickHandler(allOffers, offerCity)}
+        onClick={() => changeCityClickHandler(offers, offerCity)}
       >
         <span>{offerCity}</span>
       </a>
@@ -22,22 +22,21 @@ const CityList = (props) => {
 CityList.propTypes = {
   city: PropTypes.string.isRequired,
   changeCityClickHandler: PropTypes.func.isRequired,
-  allOffers: PropTypes.array.isRequired,
+  offers: PropTypes.array.isRequired,
   cities: PropTypes.array.isRequired
 };
 
 const mapStateToProps = (state, ownProps) =>
   Object.assign({}, ownProps, {
-    city: state.city,
-    offers: state.offers,
-    allOffers: state.allOffers,
-    cities: state.allOffers.map((offer) => offer.city.name)
+    city: state.actionUser.city,
+    offers: state.loadData.offers,
+    cities: state.loadData.offers.map((offer) => offer.city.name)
   });
 
 const mapDispatchToProps = (dispatch) => ({
-  changeCityClickHandler: (allOffers, city) => {
+  changeCityClickHandler: (offers, city) => {
     dispatch(ActionCreator.changeCity(city));
-    dispatch(ActionCreator.getOffers(allOffers, city));
+    dispatch(ActionCreator.getOffers(offers, city));
   }
 });
