+++ src/components/detail-info/detail-info.jsx
@@ -1,341 +1,333 @@
-import React, {PureComponent} from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
 
-// import ReviewsList from '../reviews-list/reviews-list';
-// import Map from '../map/map';
+import ReviewsList from '../reviews-list/reviews-list';
+import Map from '../map/map';
 import OfferList from '../offer-list/offer-list';
-import {
-  // getMapCoordinates,
-  convertRating} from '../../utils';
-import {MAX_NEARBY_OFFER} from '../../constants';
-import {Operation} from "../../reducer/reducer";
+import {getMapCoordinates, convertRating, getOtherCityOffers} from '../../utils';
+import {REVIEWS} from '../../constants';
 
-class DetailInfo extends PureComponent {
-  render() {
-    if (!this.props.allOffers.length) {
-      return null;
-    }
-    const {
-      currentOffer: {
-        isPremium,
-        price,
-        title,
-        rating,
-        photos,
-        features,
-        insideProperties,
-        hostUser
-        // reviews
-      },
-      otherOffers,
-      // activeOfferCard
-    } = this.props;
-    // const coordinates = getMapCoordinates(otherOffers, activeOfferCard);
-    return (
-      <div className="page">
-        <header className="header">
-          <div className="container">
-            <div className="header__wrapper">
-              <div className="header__left">
-                <a className="header__logo-link" href="/">
+const DetailInfo = (props) => {
+  if (!props.offers.length) {
+    return null;
+  }
+  const {
+    currentOffer: {
+      isPremium,
+      price,
+      title,
+      rating,
+      photos,
+      features,
+      insideProperties,
+      hostUser,
+      city,
+    },
+    otherOffers,
+    activeOfferCard
+  } = props;
+
+  const coordinates = getMapCoordinates(otherOffers, activeOfferCard);
+  return (
+    <div className="page">
+      <header className="header">
+        <div className="container">
+          <div className="header__wrapper">
+            <div className="header__left">
+              <a className="header__logo-link" href="/">
+                <img
+                  className="header__logo"
+                  src="/img/logo.svg"
+                  alt="6 cities logo"
+                  width="81"
+                  height="41"
+                />
+              </a>
+            </div>
+            <nav className="header__nav">
+              <ul className="header__nav-list">
+                <li className="header__nav-item user">
+                  <a
+                    className="header__nav-link header__nav-link--profile"
+                    href="#"
+                  >
+                    <div className="header__avatar-wrapper user__avatar-wrapper"></div>
+                    <span className="header__user-name user__name">
+                      Oliver.conner@gmail.com
+                    </span>
+                  </a>
+                </li>
+              </ul>
+            </nav>
+          </div>
+        </div>
+      </header>
+
+      <main className="page__main page__main--property">
+        <section className="property">
+          <div className="property__gallery-container container">
+            <div className="property__gallery">
+              {photos.map((photo, id) => (
+                <div className="property__image-wrapper" key={`photo-${id}`}>
                   <img
-                    className="header__logo"
-                    src="/img/logo.svg"
-                    alt="6 cities logo"
-                    width="81"
-                    height="41"
+                    className="property__image"
+                    src={`${photo}`}
+                    alt="Photo studio"
                   />
-                </a>
-              </div>
-              <nav className="header__nav">
-                <ul className="header__nav-list">
-                  <li className="header__nav-item user">
-                    <a
-                      className="header__nav-link header__nav-link--profile"
-                      href="#"
-                    >
-                      <div className="header__avatar-wrapper user__avatar-wrapper"></div>
-                      <span className="header__user-name user__name">
-                        Oliver.conner@gmail.com
-                      </span>
-                    </a>
-                  </li>
-                </ul>
-              </nav>
+                </div>
+              ))}
             </div>
           </div>
-        </header>
-
-        <main className="page__main page__main--property">
-          <section className="property">
-            <div className="property__gallery-container container">
-              <div className="property__gallery">
-                {photos.map((photo, id) => (
-                  <div className="property__image-wrapper" key={`photo-${id}`}>
-                    <img
-                      className="property__image"
-                      src={`${photo}`}
-                      alt="Photo studio"
-                    />
-                  </div>
-                ))}
-              </div>
-            </div>
-            <div className="property__container container">
-              <div className="property__wrapper">
-                {isPremium && (
-                  <div className="property__mark">
-                    <span>Premium</span>
-                  </div>
-                )}
-                <div className="property__name-wrapper">
-                  <h1 className="property__name">{title}</h1>
-                  <button
-                    className="property__bookmark-button button"
-                    type="button"
+          <div className="property__container container">
+            <div className="property__wrapper">
+              {isPremium && (
+                <div className="property__mark">
+                  <span>Premium</span>
+                </div>
+              )}
+              <div className="property__name-wrapper">
+                <h1 className="property__name">{title}</h1>
+                <button
+                  className="property__bookmark-button button"
+                  type="button"
+                >
+                  <svg
+                    className="property__bookmark-icon"
+                    width="31"
+                    height="33"
                   >
-                    <svg
-                      className="property__bookmark-icon"
-                      width="31"
-                      height="33"
-                    >
-                      <use xlinkHref="#icon-bookmark"></use>
-                    </svg>
-                    <span className="visually-hidden">To bookmarks</span>
-                  </button>
+                    <use xlinkHref="#icon-bookmark"></use>
+                  </svg>
+                  <span className="visually-hidden">To bookmarks</span>
+                </button>
+              </div>
+              <div className="property__rating rating">
+                <div className="property__stars rating__stars">
+                  <span style={{width: convertRating(rating) + `%`}}></span>
+                  <span className="visually-hidden">Rating</span>
                 </div>
-                <div className="property__rating rating">
-                  <div className="property__stars rating__stars">
-                    <span style={{width: convertRating(rating) + `%`}}></span>
-                    <span className="visually-hidden">Rating</span>
-                  </div>
-                  <span className="property__rating-value rating__value">
-                    {rating}
-                  </span>
-                </div>
-                <ul className="property__features">
-                  {features.map((feature, id) => (
+                <span className="property__rating-value rating__value">
+                  {rating}
+                </span>
+              </div>
+              <ul className="property__features">
+                {features.map((feature, id) => (
+                  <li
+                    className="property__feature property__feature--entire"
+                    key={`feature-${id}`}
+                  >
+                    {feature}
+                  </li>
+                ))}
+              </ul>
+              <div className="property__price">
+                <b className="property__price-value">&euro;{price}</b>
+                <span className="property__price-text">&nbsp;night</span>
+              </div>
+              <div className="property__inside">
+                <h2 className="property__inside-title">What&apos;s inside</h2>
+                <ul className="property__inside-list">
+                  {insideProperties.map((insideProperty, id) => (
                     <li
-                      className="property__feature property__feature--entire"
-                      key={`feature-${id}`}
+                      className="property__inside-item"
+                      key={`inside-property-${id}`}
                     >
-                      {feature}
+                      {insideProperty}
                     </li>
                   ))}
                 </ul>
-                <div className="property__price">
-                  <b className="property__price-value">&euro;{price}</b>
-                  <span className="property__price-text">&nbsp;night</span>
-                </div>
-                <div className="property__inside">
-                  <h2 className="property__inside-title">What&apos;s inside</h2>
-                  <ul className="property__inside-list">
-                    {insideProperties.map((insideProperty, id) => (
-                      <li
-                        className="property__inside-item"
-                        key={`inside-property-${id}`}
-                      >
-                        {insideProperty}
-                      </li>
-                    ))}
-                  </ul>
-                </div>
-                <div className="property__host">
-                  <h2 className="property__host-title">Meet the host</h2>
-                  <div className="property__host-user user">
-                    <div className="property__avatar-wrapper property__avatar-wrapper--pro user__avatar-wrapper">
-                      <img
-                        className="property__avatar user__avatar"
-                        src={`../${hostUser.avatar}`}
-                        width="74"
-                        height="74"
-                        alt="Host avatar"
-                      />
-                    </div>
-                    <span className="property__user-name">{hostUser.name}</span>
-                    <span className="property__user-status">
-                      {hostUser.status ? `pro` : ``}
-                    </span>
-                  </div>
-                  <div className="property__description">
-                    A quiet cozy and picturesque that hides behind a a river by
-                    the unique lightness of Amsterdam. The building is green and
-                    from 18th century. An independent House, strategically
-                    located between Rembrand Square and National Opera, but
-                    where the bustle of the city comes to rest in this alley
-                    flowery and colorful.
-                  </div>
-                </div>
-                <section className="property__reviews reviews">
-                  <h2 className="reviews__title">
-                    Reviews &middot;{` `}
-                    <span className="reviews__amount">{0}</span>
-                  </h2>
-                  {/* <ReviewsList reviews={reviews} /> */}
-                  <form className="reviews__form form" action="#" method="post">
-                    <label
-                      className="reviews__label form__label"
-                      htmlFor="review"
-                    >
-                      Your review
-                    </label>
-                    <div className="reviews__rating-form form__rating">
-                      <input
-                        className="form__rating-input visually-hidden"
-                        name="rating"
-                        value="5"
-                        id="5-stars"
-                        type="radio"
-                      />
-                      <label
-                        htmlFor="5-stars"
-                        className="reviews__rating-label form__rating-label"
-                        title="perfect"
-                      >
-                        <svg
-                          className="form__star-image"
-                          width="37"
-                          height="33"
-                        >
-                          <use xlinkHref="#icon-star"></use>
-                        </svg>
-                      </label>
-
-                      <input
-                        className="form__rating-input visually-hidden"
-                        name="rating"
-                        value="4"
-                        id="4-stars"
-                        type="radio"
-                      />
-                      <label
-                        htmlFor="4-stars"
-                        className="reviews__rating-label form__rating-label"
-                        title="good"
-                      >
-                        <svg
-                          className="form__star-image"
-                          width="37"
-                          height="33"
-                        >
-                          <use xlinkHref="#icon-star"></use>
-                        </svg>
-                      </label>
-
-                      <input
-                        className="form__rating-input visually-hidden"
-                        name="rating"
-                        value="3"
-                        id="3-stars"
-                        type="radio"
-                      />
-                      <label
-                        htmlFor="3-stars"
-                        className="reviews__rating-label form__rating-label"
-                        title="not bad"
-                      >
-                        <svg
-                          className="form__star-image"
-                          width="37"
-                          height="33"
-                        >
-                          <use xlinkHref="#icon-star"></use>
-                        </svg>
-                      </label>
-
-                      <input
-                        className="form__rating-input visually-hidden"
-                        name="rating"
-                        value="2"
-                        id="2-stars"
-                        type="radio"
-                      />
-                      <label
-                        htmlFor="2-stars"
-                        className="reviews__rating-label form__rating-label"
-                        title="badly"
-                      >
-                        <svg
-                          className="form__star-image"
-                          width="37"
-                          height="33"
-                        >
-                          <use xlinkHref="#icon-star"></use>
-                        </svg>
-                      </label>
-
-                      <input
-                        className="form__rating-input visually-hidden"
-                        name="rating"
-                        value="1"
-                        id="1-star"
-                        type="radio"
-                      />
-                      <label
-                        htmlFor="1-star"
-                        className="reviews__rating-label form__rating-label"
-                        title="terribly"
-                      >
-                        <svg
-                          className="form__star-image"
-                          width="37"
-                          height="33"
-                        >
-                          <use xlinkHref="#icon-star"></use>
-                        </svg>
-                      </label>
-                    </div>
-                    <textarea
-                      className="reviews__textarea form__textarea"
-                      id="review"
-                      name="review"
-                      placeholder="Tell how was your stay, what you like and what can be improved"
-                    ></textarea>
-                    <div className="reviews__button-wrapper">
-                      <p className="reviews__help">
-                        To submit review please make sure to set{` `}
-                        <span className="reviews__star">rating</span> and
-                        describe your stay with at least{` `}
-                        <b className="reviews__text-amount">50 characters</b>.
-                      </p>
-                      <button
-                        className="reviews__submit form__submit button"
-                        type="submit"
-                        disabled=""
-                      >
-                        Submit
-                      </button>
-                    </div>
-                  </form>
-                </section>
               </div>
-            </div>
-            <section className="property__map map">
-              {/* <Map
-                coordinates={coordinates}
-                activeCoordinate={activeOfferCard.coordinate}
-              /> */}
-            </section>
-          </section>
-          <div className="container">
-            <section className="near-places places">
-              <h2 className="near-places__title">
-                Other places in the neighbourhood
-              </h2>
-              <OfferList offers={otherOffers} isNearPlace />
-            </section>
-          </div>
-        </main>
-      </div>
-    );
-  }
+              <div className="property__host">
+                <h2 className="property__host-title">Meet the host</h2>
+                <div className="property__host-user user">
+                  <div className="property__avatar-wrapper property__avatar-wrapper--pro user__avatar-wrapper">
+                    <img
+                      className="property__avatar user__avatar"
+                      src={`../${hostUser.avatar}`}
+                      width="74"
+                      height="74"
+                      alt="Host avatar"
+                    />
+                  </div>
+                  <span className="property__user-name">{hostUser.name}</span>
+                  <span className="property__user-status">
+                    {hostUser.status ? `pro` : ``}
+                  </span>
+                </div>
+                <div className="property__description">
+                  A quiet cozy and picturesque that hides behind a a river by
+                  the unique lightness of Amsterdam. The building is green and
+                  from 18th century. An independent House, strategically
+                  located between Rembrand Square and National Opera, but
+                  where the bustle of the city comes to rest in this alley
+                  flowery and colorful.
+                </div>
+              </div>
+              <section className="property__reviews reviews">
+                <h2 className="reviews__title">
+                  Reviews &middot;{` `}
+                  <span className="reviews__amount">{REVIEWS.length}</span>
+                </h2>
+                <ReviewsList reviews={REVIEWS} />
+                <form className="reviews__form form" action="#" method="post">
+                  <label
+                    className="reviews__label form__label"
+                    htmlFor="review"
+                  >
+                    Your review
+                  </label>
+                  <div className="reviews__rating-form form__rating">
+                    <input
+                      className="form__rating-input visually-hidden"
+                      name="rating"
+                      value="5"
+                      id="5-stars"
+                      type="radio"
+                    />
+                    <label
+                      htmlFor="5-stars"
+                      className="reviews__rating-label form__rating-label"
+                      title="perfect"
+                    >
+                      <svg
+                        className="form__star-image"
+                        width="37"
+                        height="33"
+                      >
+                        <use xlinkHref="#icon-star"></use>
+                      </svg>
+                    </label>
 
-  componentDidMount() {
-    const {loadOffers} = this.props;
-    loadOffers();
-  }
-}
+                    <input
+                      className="form__rating-input visually-hidden"
+                      name="rating"
+                      value="4"
+                      id="4-stars"
+                      type="radio"
+                    />
+                    <label
+                      htmlFor="4-stars"
+                      className="reviews__rating-label form__rating-label"
+                      title="good"
+                    >
+                      <svg
+                        className="form__star-image"
+                        width="37"
+                        height="33"
+                      >
+                        <use xlinkHref="#icon-star"></use>
+                      </svg>
+                    </label>
+
+                    <input
+                      className="form__rating-input visually-hidden"
+                      name="rating"
+                      value="3"
+                      id="3-stars"
+                      type="radio"
+                    />
+                    <label
+                      htmlFor="3-stars"
+                      className="reviews__rating-label form__rating-label"
+                      title="not bad"
+                    >
+                      <svg
+                        className="form__star-image"
+                        width="37"
+                        height="33"
+                      >
+                        <use xlinkHref="#icon-star"></use>
+                      </svg>
+                    </label>
+
+                    <input
+                      className="form__rating-input visually-hidden"
+                      name="rating"
+                      value="2"
+                      id="2-stars"
+                      type="radio"
+                    />
+                    <label
+                      htmlFor="2-stars"
+                      className="reviews__rating-label form__rating-label"
+                      title="badly"
+                    >
+                      <svg
+                        className="form__star-image"
+                        width="37"
+                        height="33"
+                      >
+                        <use xlinkHref="#icon-star"></use>
+                      </svg>
+                    </label>
+
+                    <input
+                      className="form__rating-input visually-hidden"
+                      name="rating"
+                      value="1"
+                      id="1-star"
+                      type="radio"
+                    />
+                    <label
+                      htmlFor="1-star"
+                      className="reviews__rating-label form__rating-label"
+                      title="terribly"
+                    >
+                      <svg
+                        className="form__star-image"
+                        width="37"
+                        height="33"
+                      >
+                        <use xlinkHref="#icon-star"></use>
+                      </svg>
+                    </label>
+                  </div>
+                  <textarea
+                    className="reviews__textarea form__textarea"
+                    id="review"
+                    name="review"
+                    placeholder="Tell how was your stay, what you like and what can be improved"
+                  ></textarea>
+                  <div className="reviews__button-wrapper">
+                    <p className="reviews__help">
+                      To submit review please make sure to set{` `}
+                      <span className="reviews__star">rating</span> and
+                      describe your stay with at least{` `}
+                      <b className="reviews__text-amount">50 characters</b>.
+                    </p>
+                    <button
+                      className="reviews__submit form__submit button"
+                      type="submit"
+                      disabled=""
+                    >
+                      Submit
+                    </button>
+                  </div>
+                </form>
+              </section>
+            </div>
+          </div>
+          <section className="property__map map">
+            <Map
+              activeCityCoordinate={city.location}
+              coordinates={coordinates}
+              activeCoordinate={activeOfferCard.location}
+            />
+          </section>
+        </section>
+        <div className="container">
+          <section className="near-places places">
+            <h2 className="near-places__title">
+              Other places in the neighbourhood
+            </h2>
+            <OfferList offers={otherOffers} isNearPlace />
+          </section>
+        </div>
+      </main>
+    </div>
+  );
+};
 
 DetailInfo.propTypes = {
   idPath: PropTypes.string.isRequired,
@@ -344,12 +336,13 @@ DetailInfo.propTypes = {
 
 DetailInfo.propTypes = {
   currentOffer: PropTypes.shape({
+    city: PropTypes.object.isRequired,
     isPremium: PropTypes.bool.isRequired,
     img: PropTypes.string.isRequired,
     price: PropTypes.number.isRequired,
     title: PropTypes.string.isRequired,
     rating: PropTypes.number.isRequired,
-    type: PropTypes.oneOf([`apartment`, `private room`, `house`, `hotel`]),
+    type: PropTypes.oneOf([`apartment`, `room`, `house`, `hotel`]),
     photos: PropTypes.array.isRequired,
     features: PropTypes.array.isRequired,
     insideProperties: PropTypes.array.isRequired,
@@ -362,24 +355,18 @@ DetailInfo.propTypes = {
   otherOffers: PropTypes.array.isRequired,
   activeOfferCard: PropTypes.shape({
     id: PropTypes.string,
-    coordinate: PropTypes.array
+    location: PropTypes.object
   }),
-  loadOffers: PropTypes.func.isRequired,
+  offers: PropTypes.array.isRequired,
 };
 
 const mapStateToProps = (state, ownProps) => Object.assign({}, ownProps, {
-  currentOffer: state.allOffers.find((offer) => offer.id === `id${Number(ownProps.idPath)}`),
-  otherOffers: state.allOffers.filter((offer) => offer.id !== `id${Number(ownProps.idPath)}`).slice(0, MAX_NEARBY_OFFER),
-  activeOfferCard: state.activeOfferCard,
-  allOffers: state.allOffers,
-});
-
-const mapDispatchToProps = (dispatch) => ({
-  loadOffers: () => {
-    dispatch(Operation.loadOffers());
-  }
+  currentOffer: state.loadData.offers.find((offer) => offer.id === `id${Number(ownProps.idPath)}`),
+  otherOffers: getOtherCityOffers(ownProps.idPath, state.loadData.offers),
+  activeOfferCard: state.actionUser.activeOfferCard,
+  offers: state.loadData.offers,
 });
 
 export {DetailInfo};
 
-export default connect(mapStateToProps, mapDispatchToProps)(DetailInfo);
+export default connect(mapStateToProps)(DetailInfo);
