+++ src/components/map/map.test.js
@@ -19,7 +19,9 @@ jest.mock(`leaflet`, () => ({
 it(`Map correctly renders after relaunch`, () => {
   const tree = renderer
     .create(<Map
-      coordinates={[[1, 2]]}
+      coordinates={[{coordinate: [1, 2], zoom: 16}, {coordinate: [2, 3], zoom: 16}]}
+      activeCoordinate={{coordinate: [5, 6], zoom: 16}}
+      activeCityCoordinate={{coordinateCity: [1, 1], zoomCity: 13}}
     />)
     .toJSON();
