+++ src/mocks/offers.js
@@ -76,7 +76,7 @@ export const allOffers = {
     isPremium: false,
     price: 120,
     title: `Wood and stone place`,
-    type: `private room`,
+    type: `room`,
     rating: 4.0,
     coordinate: [52.369553943508, 4.85309666406198],
     photos: [`room.jpg`, `apartment-01.jpg`, `apartment-02.jpg`, `apartment-03.jpg`, `studio-01.jpg`, `apartment-01.jpg`],
@@ -292,7 +292,7 @@ export const allOffers = {
     isPremium: false,
     price: 120,
     title: `Wood and stone place`,
-    type: `private room`,
+    type: `room`,
     rating: 4.0,
     coordinate: [52.369553943508, 4.85309666406198],
     photos: [`room.jpg`, `apartment-01.jpg`, `apartment-02.jpg`, `apartment-03.jpg`, `studio-01.jpg`, `apartment-01.jpg`],
@@ -400,7 +400,7 @@ export const allOffers = {
     isPremium: false,
     price: 120,
     title: `Wood and stone place`,
-    type: `private room`,
+    type: `room`,
     rating: 4.0,
     coordinate: [52.369553943508, 4.85309666406198],
     photos: [`room.jpg`, `apartment-01.jpg`, `apartment-02.jpg`, `apartment-03.jpg`, `studio-01.jpg`, `apartment-01.jpg`],
