+++ src/operation/operation.js
@@ -1,0 +1,12 @@
+import ActionCreator from '../action-creator/action-creator';
+
+const Operation = {
+  loadOffers: () => (dispatch, _, api) => {
+    return api.get(`/hotels`)
+      .then((response) => {
+        dispatch(ActionCreator.loadOffers(response.data));
+      });
+  },
+};
+
+export default Operation;
