+++ src/reducer/reducer.test.js
@@ -1,13 +1,17 @@
 import MockAdapter from 'axios-mock-adapter';
-import {DEFAULT_OFFERS, DEFAULT_OFFER, SortNames, ActionType} from '../constants';
+
+import {DEFAULT_OFFERS, SortNames, ActionType} from '../constants';
 import {getCityOffers, sortOfferList} from '../utils';
-import {allOffers} from '../mocks/offers';
 import api from '../api';
-import {
-  ActionCreator,
-  reducer,
-  Operation,
-} from './reducer';
+import Operation from '../operation/operation';
+
+const mockApi = jest.genMockFromModule(`../api`);
+
+mockApi.get = jest.fn(() => {
+  return Promise.resolve({
+    data: [{fake: true}]
+  });
+});
 
 describe(`Business logic is correct`, () => {
   it(`Get offers is correctly`, () => {
@@ -58,240 +62,8 @@ describe(`Business logic is correct`, () => {
   }]);
 });
 
-describe(`Action creators work correctly`, () => {
-  it(`Action creator for change city returns correct action`, () => {
-    expect(ActionCreator.changeCity(`Paris`)).toEqual({
-      type: `CHANGE_CITY`,
-      payload: `Paris`
-    });
-  });
-  it(`Action creator for get offers for city returns action with fill array payload`, () => {
-    expect(ActionCreator.getOffers([{
-      id: `id0`,
-      city: {
-        name: `Amsterdam`,
-      }
-    },
-    {
-      id: `id1`,
-      city: {
-        name: `Paris`,
-      }
-    }], `Paris`)).toEqual({
-      type: `GET_OFFERS`,
-      payload: [
-        {
-          id: `id1`,
-          city: {
-            name: `Paris`,
-          }
-        }
-      ]
-    });
-  });
-  it(`Action creator for get offers for city returns action with empty array payload`, () => {
-    expect(ActionCreator.getOffers([{
-      id: `id0`,
-      city: {
-        name: `Amsterdam`,
-      }
-    },
-    {
-      id: `id1`,
-      city: {
-        name: `Paris`,
-      }
-    }], `Cologne`)).toEqual({
-      type: `GET_OFFERS`,
-      payload: []
-    });
-  });
-  it(`Action creator for get offers for city returns action with empty array payload`, () => {
-    expect(ActionCreator.getOffers([], `Cologne`)).toEqual({
-      type: `GET_OFFERS`,
-      payload: []
-    });
-  });
-  it(`Action creator for change sort name returns action with value payload`, () => {
-    expect(ActionCreator.changeSortName(`Top rated first`)).toEqual({
-      type: `CHANGE_SORT`,
-      payload: `Top rated first`
-    });
-  });
-  it(`Action creator for add active card returns action with object data payload`, () => {
-    expect(ActionCreator.changeActiveCard({
-      id: `id0`,
-      city: {
-        name: `Amsterdam`,
-      },
-      coordinate: [52.3909553943508, 4.85309666406198],
-    })).toEqual({
-      type: `ADD_ACTIVE_CARD`,
-      payload: {
-        id: `id0`,
-        city: {
-          name: `Amsterdam`,
-        },
-        coordinate: [52.3909553943508, 4.85309666406198],
-      }
-    });
-  });
-  it(`Action creator for add active card returns action with empty object payload`, () => {
-    expect(ActionCreator.changeActiveCard({})).toEqual({
-      type: `ADD_ACTIVE_CARD`,
-      payload: {}
-    });
-  });
-  it(`Action creator for remove active card returns action with empty object payload`, () => {
-    expect(ActionCreator.removeActiveCard()).toEqual({
-      type: `REMOVE_ACTIVE_CARD`,
-      payload: {}
-    });
-  });
-});
 
-describe(`Reducer work correctly`, () => {
-  it(`Reducer without additional parameters should return initial state`, () => {
-    expect(reducer(undefined, {})).toEqual({
-      city: `Amsterdam`,
-      offers: DEFAULT_OFFERS,
-      activeSortName: `Popular`,
-      activeOfferCard: {},
-      allOffers: allOffers.offers,
-      cities: allOffers.cities,
-    });
-  });
-  it(`Reducer should change city by a given value`, () => {
-    expect(reducer({
-      city: `Amsterdam`,
-      offers: DEFAULT_OFFERS,
-      activeSortName: `Popular`,
-      activeOfferCard: {},
-      allOffers: allOffers.offers,
-      cities: allOffers.cities
-    }, {
-      type: `CHANGE_CITY`,
-      payload: `Paris`
-    })).toEqual({
-      city: `Paris`,
-      offers: DEFAULT_OFFERS,
-      activeSortName: `Popular`,
-      activeOfferCard: {},
-      allOffers: allOffers.offers,
-      cities: allOffers.cities
-    });
-  });
-  it(`Reducer should get offers city by a given value`, () => {
-    expect(reducer({
-      city: `Amsterdam`,
-      offers: DEFAULT_OFFERS,
-      activeSortName: `Popular`,
-      activeOfferCard: {},
-      allOffers: allOffers.offers,
-      cities: allOffers.cities
-    }, {
-      type: `GET_OFFERS`,
-      payload: [{
-        id: `id1`,
-        city: {
-          name: `Paris`,
-        }
-      }]
-    })).toEqual({
-      city: `Amsterdam`,
-      offers: [{
-        id: `id1`,
-        city: {
-          name: `Paris`,
-        }
-      }],
-      activeSortName: `Popular`,
-      activeOfferCard: {},
-      allOffers: allOffers.offers,
-      cities: allOffers.cities
-    });
-  });
-  it(`Reducer should get offers city empty array by a given value`, () => {
-    expect(reducer({
-      city: `Amsterdam`,
-      offers: DEFAULT_OFFERS,
-      activeSortName: `Popular`,
-      activeOfferCard: {},
-      allOffers: allOffers.offers,
-      cities: allOffers.cities
-    }, {
-      type: `GET_OFFERS`,
-      payload: []
-    })).toEqual({
-      city: `Amsterdam`,
-      offers: [],
-      activeSortName: `Popular`,
-      activeOfferCard: {},
-      allOffers: allOffers.offers,
-      cities: allOffers.cities
-    });
-  });
-  it(`Reducer should change sort name by a given value`, () => {
-    expect(reducer({
-      city: `Amsterdam`,
-      offers: DEFAULT_OFFERS,
-      activeSortName: `Popular`,
-      activeOfferCard: {},
-      allOffers: allOffers.offers,
-      cities: allOffers.cities
-    }, {
-      type: `CHANGE_SORT`,
-      payload: `Top rated first`
-    })).toEqual({
-      city: `Amsterdam`,
-      offers: DEFAULT_OFFERS,
-      activeSortName: `Top rated first`,
-      activeOfferCard: {},
-      allOffers: allOffers.offers,
-      cities: allOffers.cities
-    });
-  });
-  it(`Reducer should add active card object data by a given value`, () => {
-    expect(reducer({
-      city: `Amsterdam`,
-      offers: DEFAULT_OFFERS,
-      activeSortName: `Popular`,
-      activeOfferCard: {},
-      allOffers: allOffers.offers,
-      cities: allOffers.cities
-    }, {
-      type: `ADD_ACTIVE_CARD`,
-      payload: DEFAULT_OFFER
-    })).toEqual({
-      city: `Amsterdam`,
-      offers: DEFAULT_OFFERS,
-      activeSortName: `Popular`,
-      activeOfferCard: DEFAULT_OFFER,
-      allOffers: allOffers.offers,
-      cities: allOffers.cities
-    });
-  });
-  it(`Reducer should remove active card empty by a given value`, () => {
-    expect(reducer({
-      city: `Amsterdam`,
-      offers: DEFAULT_OFFERS,
-      activeSortName: `Popular`,
-      activeOfferCard: DEFAULT_OFFER,
-      allOffers: allOffers.offers,
-      cities: allOffers.cities
-    }, {
-      type: `REMOVE_ACTIVE_CARD`,
-      payload: {}
-    })).toEqual({
-      city: `Amsterdam`,
-      offers: DEFAULT_OFFERS,
-      activeSortName: `Popular`,
-      activeOfferCard: {},
-      allOffers: allOffers.offers,
-      cities: allOffers.cities
-    });
-  });
-
+describe(`Reducer load data work correctly`, () => {
   it(`Should make a correct API call to /hotel`, () => {
     const apiMock = new MockAdapter(api);
     const dispatch = jest.fn();
@@ -301,10 +73,10 @@ describe(`Reducer work correctly`, () => {
       .onGet(`/hotels`)
       .reply(200, [{fake: true}]);
 
-    return offerLoader(dispatch)
+    return offerLoader(dispatch, jest.fn(), mockApi)
       .then(() => {
         expect(dispatch).toHaveBeenCalledTimes(1);
-        expect.toHaveBeenNthCalledWith(1, {
+        expect(dispatch).toHaveBeenNthCalledWith(1, {
           type: ActionType.LOAD_OFFERS,
           payload: [{fake: true}]
         });
