+++ src/store/store.js
@@ -3,13 +3,13 @@ import {composeWithDevTools} from 'redux-devtools-extension';
 import thunk from 'redux-thunk';
 import compose from 'recompose/compose';
 
-import {reducer} from '../reducer/reducer';
+import reducers from '../reducer/reducer';
 import {configureAPI} from '../api';
 
 const api = configureAPI((...args) => store.dispatch(...args));
 
 const store = createStore(
-    reducer,
+    reducers,
     compose(
         applyMiddleware(thunk.withExtraArgument(api)),
         composeWithDevTools()
